+++ src/main/java/pe/edu/upc/repository/ReserveRepository.java
@@ -10,9 +10,9 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.entity.Reserve;
 
 @Repository
-public interface ReserveRepository extends JpaRepository<Reserve, Integer>{
+public interface ReserveRepository extends JpaRepository<Reserve, Long>{
 	
 	@Query("select r from Reserve r join fetch r.reserveDetails dre join fetch dre.product where r.idReserve=?1")
-	Optional<Reserve> fetchByImportIdWhithImportDetailsWithProduct(int id);
+	Optional<Reserve> fetchByImportIdWhithImportDetailsWithProduct(Long id);
 
 }
